chore: renderable versions view
