+++ packages/next/src/pages/Versions/cells/Text/index.tsx
@@ -1,0 +1,6 @@
+'use client'
+import React from 'react'
+
+export const TextCell: React.FC<{ children?: React.ReactNode }> = ({ children }) => (
+  <span>{children}</span>
+)
