+++ packages/next/src/pages/Versions/columns.tsx
@@ -7,44 +7,11 @@ import type {
 } from 'payload/types'
 
 import type { Column } from '@payloadcms/ui'
-import { Pill, formatDate, SortColumn } from '@payloadcms/ui'
-import Link from 'next/link'
+import { SortColumn } from '@payloadcms/ui'
 import { I18n } from '@payloadcms/translations'
-
-type CreatedAtCellProps = {
-  config: SanitizedConfig
-  collectionConfig?: SanitizedCollectionConfig
-  globalConfig?: SanitizedGlobalConfig
-  date: string
-  versionID: string
-  docID: string
-  i18n: I18n
-}
-
-const CreatedAtCell: React.FC<CreatedAtCellProps> = ({
-  versionID,
-  docID,
-  config,
-  collectionConfig,
-  date,
-  globalConfig,
-  i18n,
-}) => {
-  const {
-    routes: { admin },
-    admin: { dateFormat },
-  } = config
-
-  let to: string
-
-  if (collectionConfig)
-    to = `${admin}/collections/${collectionConfig.slug}/${docID}/versions/${versionID}`
-  if (globalConfig) to = `${admin}/globals/${globalConfig.slug}/versions/${versionID}`
-
-  return <Link href={to}>{date && formatDate(date, dateFormat, i18n.language)}</Link>
-}
-
-const TextCell: React.FC<{ children?: React.ReactNode }> = ({ children }) => <span>{children}</span>
+import { CreatedAtCell } from './cells/CreatedAt'
+import { IDCell } from './cells/ID'
+import { AutosaveCell } from './cells/AutosaveCell'
 
 export const buildVersionColumns = ({
   config,
@@ -57,7 +24,7 @@ export const buildVersionColumns = ({
   config: SanitizedConfig
   collectionConfig?: SanitizedCollectionConfig
   globalConfig?: SanitizedGlobalConfig
-  docID?: string
+  docID?: string | number
   i18n: I18n
 }): Column[] => [
   {
@@ -66,15 +33,11 @@ export const buildVersionColumns = ({
     active: true,
     components: {
       Heading: <SortColumn label={t('general:updatedAt')} name="updatedAt" />,
-      renderCell: (row, data) => (
+      Cell: (
         <CreatedAtCell
-          config={config}
-          collectionConfig={collectionConfig}
-          date={data}
-          globalConfig={globalConfig}
-          versionID={row?.id}
           docID={docID}
-          i18n={i18n}
+          collectionSlug={collectionConfig?.slug}
+          globalSlug={globalConfig?.slug}
         />
       ),
     },
@@ -86,7 +49,7 @@ export const buildVersionColumns = ({
     active: true,
     components: {
       Heading: <SortColumn disable label={t('version:versionID')} name="id" />,
-      renderCell: (row, data) => <TextCell>{data}</TextCell>,
+      Cell: <IDCell />,
     },
     label: '',
   },
@@ -96,26 +59,7 @@ export const buildVersionColumns = ({
     active: true,
     components: {
       Heading: <SortColumn disable label={t('version:type')} name="autosave" />,
-      renderCell: (row) => (
-        <TextCell>
-          {row?.autosave && (
-            <React.Fragment>
-              <Pill>
-                Autosave
-                {t('version:autosave')}
-              </Pill>
-              &nbsp;&nbsp;
-            </React.Fragment>
-          )}
-          {row?.version._status === 'published' && (
-            <React.Fragment>
-              <Pill pillStyle="success">{t('version:published')}</Pill>
-              &nbsp;&nbsp;
-            </React.Fragment>
-          )}
-          {row?.version._status === 'draft' && <Pill>{t('version:draft')}</Pill>}
-        </TextCell>
-      ),
+      Cell: <AutosaveCell />,
     },
     label: '',
   },
