+++ packages/ui/src/exports/elements.ts
@@ -35,3 +35,4 @@ export { default as Popup } from '../elements/Popup'
 // export { useThumbnail } from '../elements/Upload'
 export { Translation } from '../elements/Translation'
 export { Tooltip } from '../elements/Tooltip'
+export { useTableCell } from '../elements/Table/TableCellProvider'
